Replace `frames` type (`f64` to `isize`)

+++ src/timeline.rs
@@ -44,7 +44,7 @@ pub struct Timecode {
     hours: isize,
     minutes: isize,
     seconds: isize,
-    frames: f64,
+    frames: isize,
     framerate: Framerate,
 }
 
@@ -65,16 +65,16 @@ macro_rules! tcode_hmsf_framerate {
 #[allow(unused_macros)]
 macro_rules! tcode_hms {
     ($h:tt:$m:tt:$s:tt) => {
-        Timecode::new($h, $m, $s, 0.0)
+        Timecode::new($h, $m, $s, 0)
     };
 }
 
 impl Timecode {
-    pub fn new(h: isize, m: isize, s: isize, f: f64) -> Self {
+    pub fn new(h: isize, m: isize, s: isize, f: isize) -> Self {
         Timecode::new_with_framerate(h, m, s, f, Framerate::Timestamp)
     }
 
-    pub fn new_with_framerate(h: isize, m: isize, s: isize, f: f64, fr: Framerate) -> Self {
+    pub fn new_with_framerate(h: isize, m: isize, s: isize, f: isize, fr: Framerate) -> Self {
         let mut t = Self {
             hours: h,
             minutes: m,
@@ -92,9 +92,9 @@ impl Timecode {
         let framerate = self.framerate.as_f64();
 
         if framerate != 0.0 {
-            while self.frames >= framerate {
+            while self.frames >= framerate as isize {
                 self.seconds += 1;
-                self.frames -= framerate;
+                self.frames -= framerate as isize;
             }
         }
 
@@ -113,9 +113,9 @@ impl Timecode {
         let framerate = self.framerate.as_f64();
 
         if framerate != 0.0 {
-            while self.frames < 0.0 {
+            while self.frames < 0 {
                 self.seconds -= 1;
-                self.frames += framerate;
+                self.frames += framerate as isize;
             }
         }
 
@@ -148,7 +148,7 @@ impl Timecode {
     // Add/Next
 
     pub fn next_frame(&mut self) {
-    	self.frames += 1.0;
+    	self.frames += 1;
     	self.correct_overflow();
     }
 
@@ -168,7 +168,7 @@ impl Timecode {
 
     pub fn add_by_duration(&mut self, d: Duration) {
         let secs = d.as_secs_f64();
-        self.frames += secs * self.framerate.as_f64();
+        self.frames += (secs * self.framerate.as_f64()) as isize;
 
         self.correct_overflow();
     }
@@ -176,7 +176,7 @@ impl Timecode {
     // Sub/Back
 
     pub fn back_frame(&mut self) {
-    	self.frames -= 1.0;
+    	self.frames -= 1;
     	self.correct_underflow();
     }
 
@@ -196,7 +196,7 @@ impl Timecode {
 
     pub fn sub_by_duration(&mut self, d: Duration) {
     	let secs = d.as_secs_f64();
-        self.frames -= secs * self.framerate.as_f64();
+        self.frames -= (secs * self.framerate.as_f64()) as isize;
 
         self.correct_underflow();
     }
